Improved the test suite for azure
