+++ scripts/test-azure.sh
@@ -9,5 +9,5 @@ echo "|   Testing (internal/cloudproviders/azure)"
 echo "--------------------------------------------"
 
 cd cloudproviders/azure/
-GOTEST_PALETTE="red,yellow,green" $EXEC -tags testing_azure . -v && cd -
+GOTEST_PALETTE="red,yellow,green" $EXEC -tags testing_azure -v ./... && cd -
